+++ src/eval.c
@@ -149,7 +149,8 @@ SEXP VALC_evaluate_recurse(
     }
   } else if(mode == 10 || mode == 999) {
     SEXP eval_res, eval_tmp;
-    int err_val = 0, eval_res_c;
+    int err_val = 0;
+    int eval_res_c = -1000;  // initialize to illegal value
     int * err_point = &err_val;
     eval_tmp = PROTECT(R_tryEval(lang, set.env, err_point));
     if(* err_point) {
@@ -205,7 +206,7 @@ SEXP VALC_evaluate_recurse(
               "\"err.msg\" attribute for validation token for argument `%s` must be a one length character vector."
             );
           }
-          err_call = ALIKEC_pad_or_quote(arg_lang, -1, -1, set);
+          err_call = ALIKEC_pad_or_quote(arg_lang, set.width, -1, set);
 
           // Need to make copy of string, modify it, and turn it back into
           // string
@@ -220,7 +221,7 @@ SEXP VALC_evaluate_recurse(
           // message attribute not defined, must construct error message based
           // on result of evaluation
 
-          err_call = ALIKEC_pad_or_quote(lang, -1, -1, set);
+          err_call = ALIKEC_pad_or_quote(lang, set.width, -1, set);
 
           char * err_str;
           char * err_tok;