+++ NEWS.md
@@ -2,8 +2,10 @@
 
 Initial release.
 
+* Internal: formatting strings longer than `nchar.max` no longer allowed
 * [#39](https://github.com/brodieG/validate/issues/39): `type_alike` return
   values structured like `alike`, doc fixes
+* [#38](https://github.com/brodieG/validate/issues/38): Run with valgrind
 * [#36](https://github.com/brodieG/validate/issues/36): Fix INTEGER C bug
 * [#34](https://github.com/brodieG/validate/issues/34): allow substitution of
   `.` symbol when part of `..`.