+++ NEWS.md
@@ -2,6 +2,10 @@
 
 Initial release.
 
+## 0.0.2
+
+Finalizing initial release.
+
 * [#40](https://github.com/brodieG/validate/issues/40): Removed `suggests`
   dependencies to ggplot, microbenchmark, and valaddin to improve travis build
   time